+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * feat: use error handler when reading source classes with dynamic properties
 * feat: handle magic `__get()` on the source side
+* feat: handle magic `__set()` on the target side
 
 ## 1.7.0